Bug Fixes, LEMControllerValues Struct
Fixed bugs, mainly bugs stemming from not saving button states to memory. This meant that functions could keep track of one button instead of 12 like its supposed to.

This was solved by introducing a hashmap of button types and booleans, which kept track of each individual buttons state. LEMControllerValues was uncommented and reformatted to grab the keys based off a button parameter, meaning the code will now be much cleaner.